requirements.txt
pytest
pyyaml
matplotlib
pandas